Refactored Data Strategy to use indexedDB
Did a full recode of how the data is being loaded/held on the client
side so not to have to request it everytime, instead it now gets stored
on the local indexDB storage. This allowed me to also refactor the
blueprints system so it will be possible in the future to store them
localy as well.